# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SyRide API",
        "description": "SyRide Ride-Sharing Platform – Full API Reference",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/4th_year_project/public",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/profile/{userId}/comments": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Post a comment on a user's profile",
                "operationId": "profileComment",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment posted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}